update doc & test
